requirements.txt
Flask
python-dotenv
celery
Flask-Mail